feat(loader): Selectively parse AIRD fragments
This commit changes the MelodyLoader to selectively parse only the
metadata section of AIRD files, and not keep the representation data in
memory all the time. This dramatically reduces the runtime memory
footprint and loading times for Capella models, especially when there
are lots of representations in a given model.