fix(exs): Only include full xmlns declarations on top-level
While the pure-Python version only emitted 'xmlns' declarations when
serializing the top-level element of a tree, the Rust version also did
so when serializing subtrees. Since the subtree serialization feature is
primarily intended for debugging in interactive sessions, this caused
unnecessary clutter in the output.

This commit therefore unifies the behavior of the two implementations by
only including 'xmlns' declarations when serializing an entire tree.